Baseline refactor
